upgrade bios to 6.83.00.18

+++ packages/ti/sysbios/examples/generic/bigtime/bigtime.cpp
@@ -386,7 +386,7 @@ void Clock::setMillenium()
 
 long Clock::getTicks()
 {
-    return ticks;
+    return (long)ticks;
 }
 
 int Clock::getId()
